Add the World class
