+++ Game/src/Game.h
@@ -1,8 +1,9 @@
 #pragma once
 
 #include "Sprites.h"
-#include "Entity.h"
 #include "Renderer.h"
+#include "Entity.h"
+#include "World.h"
 
 #include <vector>
 
@@ -19,15 +20,10 @@ private:
 	void Init();
 	void RunLoop();
 	void Close();
-
-	void UpdateLogic(float deltaTime);
 private:
 	uint32_t m_Width = 1200, m_Height = 900;
 	
 	Spritesheet m_Spritesheet;
 	Renderer m_Renderer{ m_Width, m_Height };
-
-	float m_Scroll = 0;
-
-	std::vector<Entity> m_Enitities;
+	World m_World{ Vector2{ 0, (float)m_Width }, Vector2{ 0, (float)m_Height } };
 };
